task edit

+++ database/migrations/2022_07_11_160811_create_task_subjects_table.php
@@ -15,7 +15,7 @@ class CreateTaskSubjectsTable extends Migration
     {
         Schema::create('task_subjects', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('task_category_id');
+            $table->unsignedInteger('task_category_id');
             $table->foreign('task_category_id')->references('id')->on('task_categories')->onDelete('cascade');
             $table->unsignedInteger('task_subcategory_id');
             $table->foreign('task_subcategory_id')->references('id')->on('task_sub_categories')->onDelete('cascade');
